requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
groq>=0.4.0
kaleido>=0.2.1
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
chardet>=5.0.0
